requirements.txt
ebooklib==0.18
html2text==2024.2.26
langchain==0.2.14
openai==1.12.0
python-dotenv==1.0.1
